+++ src/ValueObject/Human.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * (c) Meritoo.pl, http://www.meritoo.pl
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Meritoo\Common\ValueObject;
+
+use Meritoo\Common\Traits\ValueObject\HumanTrait;
+
+/**
+ * Human
+ *
+ * @author    Meritoo <github@meritoo.pl>
+ * @copyright Meritoo <http://www.meritoo.pl>
+ */
+class Human
+{
+    use HumanTrait;
+}